added isUsingSubWeaponX
int that determines which animations to use based on which subweapon is equipped. this is a test, if results are favorable then i will adopt this method for attacking with different weapons. may end up scrapping the sub weapon x in favor of using different animation layers for each weapon.