imported project to Tays comp in miami
messing with keyboard support. trying to find a good way to use keypresses when the game is searching for an axis value. may end up searching for keypressed and manaully updating axisq values as a seperate function? or maybe new input manager has a solution for it. will be reading on this.